// patientor/src/types.ts
import { z } from 'zod';
import { newEntrySchema } from './utils';

  export interface Diagnosis {
    code: string;
    name: string;
    latin?: string;
  }

  interface BaseEntry {
    id: string;
    description: string;
    date: string;
    specialist: string;
  
    diagnosisCodes?: Array<Diagnosis['code']>;
  }
  
  export enum HealthCheckRating {
    "Healthy" = 0,
    "LowRisk" = 1,
    "HighRisk" = 2,
    "CriticalRisk" = 3
  }
  
  interface HealthCheckEntry extends BaseEntry {
    type: "HealthCheck";
    healthCheckRating: HealthCheckRating;
  }
  
  interface HospitalEntry extends BaseEntry {
    type: "Hospital";
    discharge: {
      date: string;
      criteria: string;
    }
  }
  
  interface OccupationalHealthcareEntry extends BaseEntry {
    type: "OccupationalHealthcare";
    employerName: string;
    sickLeave?: {
      startDate: string;
      endDate: string;
    }
  
  }
  export type Entry =
    | HospitalEntry
    | OccupationalHealthcareEntry
    | HealthCheckEntry;
  
  
  export interface Patient {
    id: string;
    name: string;
    occupation: string;
    gender: string;
    ssn?: string;
    dateOfBirth?: string;
    entries: Entry[];
  }

  export enum Gender {
    Male = "male",
    Female = "female",
    Other = "other"
  }

  export type PatientFormValues = Omit<Patient, "id" | "entries" | "ssn">;

  export type NewPatient = Omit<Patient, "id">;

  export type NonSensitivePatient = Omit<Patient, 'ssn' | 'entries'>;

  export type NewPatientEntry = z.infer<typeof newEntrySchema>;